feat(video): pass mime type to the service
